Emit IR for recursively-defined types
This is made possible by treating any non-primitive (and non-String) type identifier T as T*,
and registering T as an opaque struct type as soon as the TypeBuilder encounters it (allowing
for recursive field-building to complete).

After resolution, Neeilang continues to treat a (likely heap-allocated) T via its transparent pointer T*
(that is, only fixed-size primitives will ever be stack-allocated).