Generate IR for methods without relying on hacky type-table
Previously, func names were mapped (globally) to Type objects in the
TypeTable. Because this table was not scoped according to the same rules
as other front-end passes, methods and functions with the same name could
override each other.
Now, methods for a particular type, along with it's fields, are stored
in that type's Type struct, and this allows methods of multiple classes
to have the same name (and even have the name of a global function).