Define memory layout for objects (add a header) in LLVM backend
Previously, generated code for objects laid them out as 'C-structs'
in memory. In this commit, a transparent (to most of the code generator)
object header is introduced.

This is a prerequisite to implement features like OOP (such as virtual
functions via a VTable ptr in the header) and garbage collection (via
a dedicated GC tracing byte that may be bitmasked for tri-colour
marking).

The header defined in this commit contains the GC byte and a VTable
pointer, but this will almost certainly change; the more important goal
for this commit was to adjust existing field-access codegen to keep
functinal tests (./run_functional_tests.sh) running.