[llvm-backend] Implement virtual functions
This commit implements virtual method calling in a fairly orthodox way. VTables
are built for each type, and the vtable pointer field in the object header is used
to access the vtable for dynamic dispatch.
In order to get this to work, the commit also changes object layout (object.h) and
the LLVM type building algorithm.