[frontend] Modify array type syntax to specify dimensions
Previously, the syntax for specifying array types was as follows:

[ Type ]   for 1D arrays
[[ Type ]] for 2D arrays etc.

In this commit, it has been modified to:

Type[size] for fixed-sized arrays, and
Type[]     for unknown-sized arrays,

where size is an arbitrary Int expression.

The disadvantage of the old syntax lies in the lack of easy dimension
specifiers; this quirk forces implementation of dynamic-sized arrays
(the only ones available in the language under that scheme) at code-generation.

Codegen for fixed-size arrays is much simpler, and variable-sized
arrays can be implemented in source on top of fixed-size arrays.