[frontend/llvm backend] Enable polymorphism in function args + return
Object pointers have to be properly bitcasted to enable passing of
subclasses as superclasses to functions (and as return values), as
LLVM IR is strongly typed.

In addition, this commit also fixes a bug in superclass field access
from a subclass (as the field_idx method only returned the index
within the current class and not the object's entire header).